+++ pkg/querier/tripperware/queryrange/limits_test.go
@@ -236,6 +236,7 @@ type mockLimits struct {
 	maxQueryLookback       time.Duration
 	maxQueryLength         time.Duration
 	maxCacheFreshness      time.Duration
+	maxQueryResponseSize   int64
 	queryVerticalShardSize int
 }
 
@@ -253,6 +254,10 @@ func (mockLimits) MaxQueryParallelism(string) int {
 
 func (m mockLimits) MaxCacheFreshness(string) time.Duration {
 	return m.maxCacheFreshness
+}
+
+func (m mockLimits) MaxQueryResponseSize(string) int64 {
+	return m.maxQueryResponseSize
 }
 
 func (m mockLimits) QueryVerticalShardSize(userID string) int {